Fix dtype mismatches in Column construction and address PR feedback
- Fix quantile to use actual dtype from plc result (always FLOAT64)
- Fix is_month_end to use from_pylibcudf for correct dtype inference
- Fix DatetimeTZColumn._get_dt_field to infer dtype then cast to int32
- Update test expectation for hour attribute to int32
- Remove redundant is_decimal256 check (handled by is_decimal branch)

+++ python/cudf/cudf/tests/indexes/datetimeindex/test_attributes.py
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: Copyright (c) 2025, NVIDIA CORPORATION.
+# SPDX-FileCopyrightText: Copyright (c) 2025-2026, NVIDIA CORPORATION.
 # SPDX-License-Identifier: Apache-2.0
 
 import numpy as np
@@ -43,7 +43,8 @@ def test_tz_aware_attributes_local():
     ]
     dti = cudf.DatetimeIndex(data).tz_localize("UTC").tz_convert("US/Eastern")
     result = dti.hour
-    expected = cudf.Index([9, 9, 9], dtype="int16")
+    # Hour is always widened to int32 (from int16) for both tz-aware and tz-naive
+    expected = cudf.Index([9, 9, 9], dtype="int32")
     assert_eq(result, expected)
